+++ common/src/main/java/store/_0982/common/kafka/dto/OrderEvent.java
@@ -11,14 +11,14 @@ import java.util.UUID;
 public class OrderEvent extends BaseEvent {
     private final UUID id;
     private final UUID memberId;
-    private final int price;
+    private final long price;
     private final int quantity;
     private final String address;
     private final String addressDetail;
     private final String postalCode;
     private final String status;    // Order의 status
 
-    public OrderEvent(Clock clock, UUID id, UUID memberId, int price, int quantity, String address,
+    public OrderEvent(Clock clock, UUID id, UUID memberId, long price, int quantity, String address,
                       String addressDetail, String postalCode, String status) {
         super(clock);
         this.id = id;
